refactor🧹 : Recoil 코드 정리 및 fetch 부분 react query로 정리
